requirements.txt
# Core LangChain e extensões
langchain
langchain-community
langchain-ollama
langchain-openai
langchain-huggingface

# LLMs locais
torch  # Usado por transformers e modelos locais
tiktoken

# Processamento de texto e PDF
unstructured
PyPDF2
pymupdf
pdfminer.six
python-docx
openpyxl
xlrd


#Recall
sentence-transformers 
scikit-learn 
rouge-score

# Imagem
pillow

# Infraestrutura
faiss-cpu
tqdm
pydantic
python-dotenv

# Interface web
streamlit

# Desenvolvimento e testes
pytest




#langchain
#langchain-core
#langchain-community
#langchain-huggingface
#streamlit
#tqdm
#pydantic
#python-dotenv
#unstructured
#pdfminer.six
#faiss-cpu
#python-docx
#pytest
#tiktoken
##llama-cpp-python
#langchain_ollama 
#langchain_openai 
#langchain-ollama
#openpyxl
#xlrd
#torch
#PyPDF2